Improved search function Part 9 🦭
