+++ rs/wasm/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wasm-book"
-version = "0.8.0"
+version = "0.9.0"
 authors = ["CoralPink <teqt6ytqt@mozmail.com>"]
 edition = "2024"
 rust-version = "1.85"
@@ -21,6 +21,7 @@ opt-level = "s"
 strip = "symbols"
 
 [dependencies]
+arrayvec = "0.7.6"
 macros = { path = "macros" }
 serde = { version = "=1.0.219", features = ["derive"] }
 serde-wasm-bindgen = "=0.6.5"
